Change to a more obvious language-select icon
We've been using ye olde globe icon for language selection, but I don't believe it's a good icon for this. The globe represents the Internet, not languages. And even if it's meant to be the world, countries do not map to language.

I've left the globe-icon include in the repo in case future users do prefer to use that.