Avoid compressing blanklines globally, use Liquid strip instead
For valid XML in package.opf, we cannot have a blank line at the
start of the file. Previously, on epub output we removed blank lines
globally with the compress layout. This was overkill, and removes
linebreaks that we need to retain in CDATA blocks, which kramdown
uses for TeX that contains line breaks. This change removes the
blanklines setting from compress, and instead uses a Liquid capture
with a strip filter to remove the blank line in package.opf.