Workaround para indent issue introduced in #416
In #416 I created a dedicated `$paragraph-indent` variable,
which seems a better way to control paragraph indentation
than just using `$line-height-default` (which tends to be
too great an indent visually). However, there are far more
places in the CSS that use `$line-height-default` for indentation
(e.g. subsequent paras and list indentation) that also need
updating. So, for now, setting `$paragraph-indent` to
`$line-height-default` keeps the old behaviour while I work on
a deeper refactoring in the many partials that handle indentation.