Fix faulty check for whether we're running Prince
While our previous check did work in Prince, it threw an error
if we opened the page in, say, a browser where the Prince object
does not exist. This made it difficult to debug PDF output in a browser.
We can also remove the check for Phantom, since the Prince object
would not exists there, either.