Fix links to translated images in live builds
This makes the check for whether we're using external media less strict.
Previously, not having a `live` setting for external media, while
actually using an external media repo, led the metadata include to believe
that the translation should use the parent language's images.
Now, if *any* external media setting is used, the metadata must assume
that translated images could exist, and link to them.

This is related to the changes in db222d0808cc8f9612543ce515fcefb820f00261