Remove unnecessary links in epub template
Discovered when demoing recently, meta.yml includes
some values in the epub toc that are very specific to
the dummy text in the template. The slightest change to
the dummy text headings will cause invalid epub output,
which will frustrate and confuse new users, who can't be
expected to know to fix meta.yml. This removes that
unnecessary little complexity.