Exit the output script if Jekyll builds fail
Till now, if there is an error in a Jekyll build
(e.g. if Sass or YAML are invalid), the output script
will just skip that error and generate a book/site
from the HTML ni the _site folder from a previous build.
If you miss that the Jekyll build failed, you can end up
looking at the wrong output with awkward consequences.

This change forces the output script to exit completely
if a Jekyll build fails. Note that it does not exit
if a Jekyll serve fails, since those errors are easier to spot
at the command line, and can be fixed live.