Fix clicking between bookmarks in different accordion sections on same page
With accordions on, we listen for clicks on the page
so that we can look up which accordion section a target is in
and open it before scrolling there. Our accordions script
was not listening for these clicks in the bookmarks modal.
This change adds listeners to links in the bookmarks modal.