Use store for files list when rendering index
Previously, we parsed the array of files from files-listed.html
separately, and differently, when creating the store and the index.
In some cases, the store and idnex could become mismatched,
resulting in incorrect search results. This fix has the
render-search-index script use the store for its file list,
to ensure that the two are identical.