Temporarily disable kramdown v2 dependency
This will be resolved in #509. For now, if we try to use
kramdown v2 dependencies with Jekyll 3.8 or lower, bundler
will error because kramdown-parser-gfm depends on kramdown v2.
This error breaks builds on Netlify. GH Pages will build
either way because it ignore the Gemfile and currently uses
Jekyll 3.9 anyway.