Add font-style management
When we inject SVGs, those SVGs' style attributes often contain fonts that are named differently from how CSS names them. For instance, the SVG might include 'OpenSans-Bold', but we need that to be 'Open Sans' with a separate attribute for 'bold', in order for our CSS to apply the correct font.

These mappings are managed in svg-management.js. We've long managed font names and weights. This change adds the ability to manage font styles, too.