Remove _items functionality
While it's a pity to lose the ability to store repeatable content
in _items, there are good reasons to remove it:

- over time, we have not seen this feature used
- _items adds to build times even when not used
- includes in items do not work reliably, in particular
  figures in questions to do render, probably due to the order
  in which Jekyll expands Liquid tags

Repeatable items can still be added in _includes if necessary.

I've not removed _items from run-* scripts, because we expect
to replace those scripts with a single cross-platform run.js soon.