'-ah-flush-zone' and '-ah-adjust-last-line-spacing'.
'-ah-flush-zone' nudges the last line of justified text to fill the
line if unadjusted end of the line is within the 'flush zone'.

'-ah-adjust-last-line-spacing' adjusts the spacing of the last line of
justified text so there's less visual difference between a possible
widely-space second last line and an unjustified last line with normal
spacing between words.