Also highlight search-query-stems in pages linked from search
When we search, elasticlunr actually searches for the 'stems'
of the words in our query. E.g. if we search for 'processing'
elasticlunr searches the index for 'process' and therefore
finds terms like 'process', 'processed' and 'processor'.

This is smart. But when we click through to a search result,
mark.js currently only highlights on the page our original
search query, 'processing'. If the page only contains the
word 'processed', nothing will be highlighted, which is very
confusing. To the user, the search highlighting seems broken.

This change fixes that by adding an extra URL parameter to
URLs on search-result pages: 'search_stem'. And it makes mark.js
also highlight those search stems in addition to the original
search query. This way, users will more easily find where on
the page elasticlunr matched a phrase to their search query.