Add refresh-indexes command
These changes introduce an important fix: till now, it was possible
to generate a search index with docs enabled, and then use that index
for a live deployment with docs disabled. This could mean that users
using search got results that include the docs pages, even though
those docs pages are not part of the site they're on.

We do want the docs to be included in search for development purposes,
especially on the default EBT itself.

To avoid this problem, we need to have two search indexes:
one listing the docs and one not. That's what this work achieves.
When generating search indexes, we build both.
The search-notifications.js now loads the correct search index,
depending whether the docs are currently enabled or not.