Remove bad legacy CSS properties
This removes some short-sighted code I added years ago
that was meant to allow using bold or italic markdown
spans for loosening or tightening inline text. We've
never done this in practice because it's a terrible,
unsemantic idea.

It's come to light now because `font-weight: inherit`
applies to a paragraph with a font-weight of, say, 300,
inherits the font-weight of its container, which has
a default font-weight of 400/normal.