Use more robust test before loading lazy images
For some unknown reason, when indexing tags are in a particular position (e.g. before the first `h1`), the `intersectionObserver` was not seeing images as `isIntersecting` after `DOMContentLoaded`. So, instead, I'm now using a `mutationObserver` as we already do for the accordion and bookmarking. Perhaps the changes that index tags make to the DOM were interfering with the `insersectionObserver`'s view of the document.