Provide format-specific .thumbnail figures
In real-world projects, we often find that a 'thumbnail'
in one format does not work well in another format.
This lets us make a figure a thumbnail on a format-by-format
basis, and use other (often custom) classes in addition
as needed.

I have my doubts about whether the template should even have
a `thumbnail` class by default, since it's an opinionated
design feature. But we use it so often that a default seems
worth having.