Load CSS as a Prince user style sheet to support SVG styling
Since Prince 14 (including Prince for Books) SVGs referenced
with <img src""> do not have access to the parent doc's CSS.
This means they don't get the parent doc's fonts, which we
have depended on. It will be much more onerous and fragile to
try to have each SVG load its own fonts. So by loading the CSS
as a user style sheet like this, Prince will apply the CSS
to all documents including SVGs referenced with <img src="">.

I've also added some failsafes and simpler logging so that
Prince problems are easier to spot in the output logs.