Remove fallback to avoid invalid epubs
The fallback for the toc was set to the default-start-page,
which by default is the cover. But the cover has no
'nav' element for EPUBCheck to validate. So if the contents-page
was not set, and if the cover was processed by the epub-package
include before the file that contains the nav and is set in
default.yml as the `filename: "toc"`, then epub-package
incorrectly thinks that the cover file contains a `nav` element.
This leads to package.opf declaring two file with `nav`
elements -- cover and the real `nav` file -- which causes
EPUBCheck to throw errors.