Do not process PDF HTML with Cheerio in XML mode
While Cheerio's XML mode is necessary for epub output,
I can't remember why we're using it for PDF output.
Hopefully it is not important, so we should
keep an eye on this. What is a problem is that when
a paragraph contains more than about two lines
that end in linebreak elements (<br/>), Cheerio
(version 0.*) incorrectly adds extra linebreaks
to the end of the paragraph.