Replace search-store with YAML-based file lists
The search-store depended on Jekyll being built
in order to generate the search-store required
to build with Jekyll. This was clumsy and
non-intuitive. It was also easy to implement, but
could result in files being indexed when they
are not intended to be part of a publication.
This new approach relies on _data/works more heavily,
and specifically on file lists. Only files
in file lists, files in nav.yml, and docs files
(where docs are enabled in config) will be indexed.