Clarify how to enable/disable collections
The aim here is to avoid having Jekyll process
collections that we don't need for a given format,
so that build times are shorter.

Turning on and off multiple collections is
a quite poorly and confusingly documented area
in Jekyll. So my code comments and template
configs here are based on forum discussions and
my own testing.