Auto handle named entities in divs
Till now, named entities like &nbsp; could not be
used in tables inside figures. This is because
our process couldn't re-encode these named entities
as numeric ones, since kramdown doesn't 'reach'
inside HTML islands. And we need these entities to
be valid in XML for epub output, and for processing
steps like Chrrio not to break entities in PDF output.

So this adds a step to replace named entities with
numeric entities after Jekyll runs, before other processing.

We act on the HTML as a string,
and do not try to parse it, because tools that parse
the HTML into an AST like parse5, jsdom, or htmlparser2
break XML validity and double-encode entities
when they render the HTML back to us.

This processing is not necessary for web and app outputs.